+++ src/display.rs
@@ -22,6 +22,7 @@ pub enum DisplayType {
     Classic(DisplayFilter),
     NvimNotify(DisplayFilter),
     VirtualText(DisplayFilter),
+    VirtualLine(DisplayFilter),
     Terminal(DisplayFilter),
     TerminalWithCode(DisplayFilter),
     LongTempFloatingWindow(DisplayFilter),
@@ -47,6 +48,7 @@ impl FromStr for DisplayType {
         match display_type {
             "Classic" => Ok(Classic(display_filter)),
             "VirtualText" => Ok(VirtualText(display_filter)),
+            "VirtualLine" => Ok(VirtualLine(display_filter)),
             "Terminal" => Ok(Terminal(display_filter)),
             "TerminalWithCode" => Ok(TerminalWithCode(display_filter)),
             "LongTempFloatingWindow" => Ok(LongTempFloatingWindow(display_filter)),
@@ -76,6 +78,7 @@ impl fmt::Display for DisplayType {
         let name = match &self {
             DisplayType::Classic(filter) => "Classic".to_string() + &filter.to_string(),
             DisplayType::VirtualText(filter) => "VirtualText".to_string() + &filter.to_string(),
+            DisplayType::VirtualLine(filter) => "VirtualLine".to_string() + &filter.to_string(),
             DisplayType::Terminal(filter) => "Terminal".to_string() + &filter.to_string(),
             DisplayType::TerminalWithCode(filter) => {
                 "TerminalWithCode".to_string() + &filter.to_string()
@@ -115,6 +118,7 @@ pub fn display(result: Result<String, SniprunError>, nvim: Arc<Mutex<Neovim>>, d
                 return_message_classic(&result, &nvim, &data.return_message_type, data, *f)
             }
             VirtualText(f) => display_virtual_text(&result, &nvim, data, *f),
+            VirtualLine(f) => display_virtual_line(&result, &nvim, data, *f),
             Terminal(f) => display_terminal(&result, &nvim, data, *f),
             TerminalWithCode(f) => display_terminal_with_code(&result, &nvim, data, *f),
             LongTempFloatingWindow(f) => display_floating_window(&result, &nvim, data, true, *f),
@@ -177,6 +181,85 @@ pub fn send_api(
     }
 }
 
+pub fn display_virtual_line(
+    result: &Result<String, SniprunError>,
+    nvim: &Arc<Mutex<Neovim>>,
+    data: &DataHolder,
+    filter: DisplayFilter,
+) {
+    info!("range is : {:?}", data.range);
+    let namespace_id = nvim.lock().unwrap().create_namespace("sniprun").unwrap();
+    if (filter == OnlyOk) && result.is_err() || (filter == OnlyErr) && result.is_ok() {
+        if let Err(SniprunError::InterpreterLimitationError(_)) = result {
+            return; // without clearing the line
+        }
+        // clear the current line
+        let last_line = data.range[1] - 1;
+        let _ = nvim.lock().unwrap().command(&format!(
+            "call nvim_buf_clear_namespace(0,{},{},{})",
+            namespace_id,
+            data.range[0] - 1,
+            last_line + 1
+        ));
+
+        return; //don't display unasked-for things
+    }
+
+    info!("namespace_id = {:?}", namespace_id);
+
+    let last_line = data.range[1] - 1;
+    let res = nvim.lock().unwrap().command(&format!(
+        "call nvim_buf_clear_namespace(0,{},{},{})",
+        namespace_id,
+        data.range[0] - 1,
+        last_line + 1
+    ));
+    info!("cleared previous virtual_line? {:?}", res);
+
+    let hl_ok = "SniprunVirtualTextOk";
+    let hl_err = "SniprunVirtualTextErr";
+    let res = match (result, filter) {
+        (Ok(message_ok), OnlyOk) | (Ok(message_ok), Both) => {
+            if no_output_wrap(message_ok, data, &DisplayType::VirtualLine(filter)).is_empty() {
+                return;
+            }
+            nvim.lock().unwrap().command(&format!(
+                "lua require\"sniprun.display\".display_virt_line({},{},\"{}\",\"{}\")",
+                namespace_id,
+                last_line,
+                &no_output_wrap(message_ok, data, &DisplayType::VirtualLine(filter))
+                    .replace('\n', "\\\n"),
+                hl_ok
+            ))
+        }
+        (Err(message_err), OnlyErr) | (Err(message_err), Both) => {
+            if no_output_wrap(
+                &message_err.to_string(),
+                data,
+                &DisplayType::VirtualLine(filter),
+            )
+            .is_empty()
+            {
+                return;
+            }
+            nvim.lock().unwrap().command(&format!(
+                "lua require\"sniprun.display\".display_virt_line({},{},\"{}\",\"{}\")",
+                namespace_id,
+                last_line,
+                &no_output_wrap(
+                    &message_err.to_string(),
+                    data,
+                    &DisplayType::VirtualLine(filter)
+                )
+                .replace('\n', "\\\n"),
+                hl_err
+            ))
+        }
+        _ => Ok(()),
+    };
+    info!("done displaying virtual lines, {:?}", res);
+}
+
 pub fn display_virtual_text(
     result: &Result<String, SniprunError>,
     nvim: &Arc<Mutex<Neovim>>,
@@ -226,7 +309,7 @@ pub fn display_virtual_text(
                 return;
             }
             nvim.lock().unwrap().command(&format!(
-                "lua require\"sniprun.display\".display_extmark({},{},\"{}\",\"{}\")",
+                "lua require\"sniprun.display\".display_virt_text({},{},\"{}\",\"{}\")",
                 namespace_id,
                 last_line,
                 shorten_ok(&no_output_wrap(
@@ -248,7 +331,7 @@ pub fn display_virtual_text(
                 return;
             }
             nvim.lock().unwrap().command(&format!(
-                "lua require\"sniprun.display\".display_extmark({},{},\"{}\",\"{}\")",
+                "lua require\"sniprun.display\".display_virt_text({},{},\"{}\",\"{}\")",
                 namespace_id,
                 last_line,
                 shorten_err(&no_output_wrap(