+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.3.17
 - better fallback messages
 - Terminal display mode fixes (courtesy of @dbeecham)
+- VirtualLine display mode
 
 ## v1.3.16
 - Fix nvim-notify timeout and add render style option